requirements.txt
aiofiles==0.4.0
asgiref==3.3.1
boto3==1.17.27
botocore==1.20.27
certifi==2017.7.27.1
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.3
django-cors-headers==3.7.0
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-storages==1.10.1
djangorestframework==3.12.2
future==0.18.2
gunicorn==20.0.4
idna==2.7
Jinja2==2.10.3
jmespath==0.10.0
MarkupSafe==1.1.1
Paperboy==1.0.1
Pillow==8.0.1
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2017.2
requests==2.20.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
starlette==0.12.10
urllib3==1.26.2
whitenoise==5.2.0